+++ sql/snowflake_ddl/facts/fact_transaction.sql
@@ -1,6 +1,6 @@
 CREATE OR REPLACE TABLE  FACT_TRANSACTION( 
     transaction_id INT primary key,
-    wallet_id INT NOT NULL,
+    wallet_id BIGINT NOT NULL,
     transaction_type_id INT NOT NULL,
     transaction_amount DECIMAL(18,2) NOT NULL,
     transaction_status VARCHAR(20) NOT NULL,
